Change averageTime from int to float
It's a major change for existing projects (they might need to adapt their code) but it is more close to what we are used to get from a ping.

+++ ESP32Ping.h
@@ -52,7 +52,7 @@ public:
 
     bool ping(const char *host, byte count = 5);
 
-    int averageTime();
+    float averageTime();
 
 protected:
     static void _ping_sent_cb(void *opt, void *pdata);
@@ -63,7 +63,7 @@ protected:
     ping_option _options;
 
     static byte _expected_count, _errors, _success;
-    static int _avg_time;
+    static float _avg_time;
 };
